moved some stuff from the ThoughtsFeed component to a new ThoughtCard comp
