+++ code/src/components/ThoughtsFeed/ThoughtsFeed.jsx
@@ -1,9 +1,8 @@
 /* eslint-disable no-underscore-dangle */
 import React, { useState, useEffect } from 'react';
-import moment from 'moment';
 
 import ThoughtForm from 'components/ThoughtForm/ThoughtForm';
-import LikeButton from '../LikeButton/LikeButton';
+import ThoughtCard from 'components/ThoughtCard/ThoughtCard';
 import styles from './ThoughtsFeed.module.css';
 
 const ThoughtsFeed = () => {
@@ -27,36 +26,10 @@ const ThoughtsFeed = () => {
     return <h2 className={styles.loadingMessage}>Loading...</h2>
   }
 
-  const handleMessageLiked = (likedThoughtId) => {
-    const updatedThoughtsFeed = thoughtsFeed.map((thought) => {
-      if (thought._id === likedThoughtId) {
-        thought.hearts += 1
-      }
-      return thought
-    })
-    setThoughtsFeed(updatedThoughtsFeed);
-  };
-
   return (
     <section className={styles.feedGrid}>
       <ThoughtForm setThoughtsFeed={setThoughtsFeed} />
-      {thoughtsFeed.map((thought) => (
-        <div key={thought._id} className={styles.thoughtCard}>
-          <h4 className={styles.thoughtMessage}>{thought.message}</h4>
-          <div className={styles.likesAndTimeContainer}>
-            <div className={styles.likesContainer}>
-              <LikeButton
-                thoughtId={thought._id}
-                thought={thought}
-                handleMessageLiked={handleMessageLiked} />
-              <p className={styles.amountOfLikes}> x {thought.hearts}</p>
-            </div>
-            <p className={styles.timePosted}>
-              {moment(thought.createdAt).fromNow()}
-            </p>
-          </div>
-        </div>
-      ))}
+      <ThoughtCard thoughtsFeed={thoughtsFeed} setThoughtsFeed={setThoughtsFeed} />
     </section>
   )
 };
