+++ code/src/components/ThoughtsFeed/ThoughtsFeed.module.css
@@ -7,44 +7,6 @@
   max-width: 700px;
 }
 
-.thoughtCard {
-  background-color: white;
-  border: 1px solid black;
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  box-shadow: 0.3rem 0.3rem black;
-  padding: 1.5rem;
-}
-
-.thoughtMessage {
-  font-size: 1rem;
-  font-weight: 400;
-}
-
-.likesContainer {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.likesAndTimeContainer {
-  display: flex;
-  align-items: center;
-  color: gray;
-  font-size: 0.8rem;
-  justify-content: space-between;
-}
-
-.amountOfLikes {
-}
-
-.timePosted {
-}
-
-.loadingMessage {
-}
-
 @media (max-width: 705px) {
   .feedGrid {
     width: 90%;
